pyproject.toml
[project]
name = "fuel-route"
version = "0.1.0"
description = "Fuel-optimal route planner API built with Django"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "requests>=2.32.3",
    "shapely>=2.0.6",
    "geopy>=2.4.1",
    "python-dotenv>=1.0.1",
    "staticmap>=0.5.7",
    "Pillow>=11.0.0",
    "django-apscheduler>=0.6.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fuel_route", "route_optimizer"]
